requirements.txt
aiohttp>=3.9.5
aiosignal>=1.3.1
asgiref>=3.8.1
attrs>=23.2.0
discord>=2.3.2
Django>=5.0.6
frozenlist>=1.4.1
idna>=3.7
multidict>=6.0.5
pillow>=10.3.0
sqlparse>=0.5.0
tzdata>=2024.1
yarl>=1.9.4
google-auth>=2.3.0
keyring
praw